# api/swagger/swagger.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API voyage. Documentation",
        "description": "Cette API permet aux tours-opérateurs de gérer les avis de leurs clients sur les voyages qu'ils proposent.\r\n\r\n    ",
        "contact": {
            "name": "the developer",
            "email": "sengerromain75@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "https://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1/php/agence_voyage/api/",
            "description": "Serveur local de développement"
        }
    ],
    "paths": {
...
        "/client/create": {
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Ajouter un client",
                "description": "Ajouter un client",
                "operationId": "856d9fa4a49e5034542af838f77fb6e8",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "prenom",
                                    "nom",
                                    "email"
                                ],
                                "properties": {
                                    "prenom": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "John.Doe@gmail.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inserer les datas"
                    },
                    "400": {
                        "description": "Tous les champs sont obligatoires"
                    },
                    "405": {
                        "description": "Méthode non autorisée"
                    }
                }
            }
        },
        "/client/read/{clientID}": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Voir un client",
                "description": "Afficher un client",
                "operationId": "f0fa58d6ed8e44256fdb76a325b01210",
                "parameters": [
                    {
                        "name": "clientID",
                        "in": "path",
                        "description": "clientID",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Client trouvé",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Client"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Client non-trouvé"
                    },
                    "405": {
                        "description": "Méthode non-autorisé"
                    }
                }
            }
        },
        "/client/readAll": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Afficher tous les clients",
                "description": "Afficher tous les clients",
                "operationId": "b7cf2c4bcbe496d142f4d212ba5ba956",
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Client"
                                    }
                                }
                            }
                        }
                    },
                    "405": {
                        "description": "Mauvaise Méthode"
                    }
                }
            }
        },
        "/client/update": {
            "put": {
                "tags": [
                    "Client"
                ],
                "summary": "Modifier un client",
                "description": "Modifier un client",
                "operationId": "80e1c0b753a851e1899d183d66fddb1f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "clientID",
                                    "prenom",
                                    "nom",
                                    "email"
                                ],
                                "properties": {
                                    "prenom": {
                                        "type": "string",
                                        "example": "John"
                                    },
                                    "nom": {
                                        "type": "string",
                                        "example": "Doe"
                                    },
                                    "email": {
                                        "type": "string",
                                        "example": "John.Doe@gmail.com"
                                    },
                                    "clientID": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Inserer les datas"
                    },
                    "400": {
                        "description": "Tous les champs sont obligatoires"
                    },
                    "405": {
                        "description": "Méthode non autorisée"
                    }
                }
            }
        },
        "/client/delete": {
            "delete": {
                "tags": [
                    "Client"
                ],
                "summary": "Supprimer un client",
                "description": "Supprimer un client",
                "operationId": "49c1c29cf36139739f5945e43da1c5d9",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "clientID"
                                ],
                                "properties": {
                                    "clientID": {
                                        "description": "ID du client à supprimer",
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Supprimer les datas"
                    },
                    "400": {
                        "description": "Tous les champs sont obligatoires"
                    },
                    "405": {
                        "description": "Méthode non autorisée"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Client": {
                "description": "Représente un client",
                "properties": {
                    "clientID": {
                        "description": "Clé unique du client",
                        "type": "integer"
                    },
                    "prenom": {
                        "description": "Prénom du client",
                        "type": "string"
                    },
                    "nom": {
                        "description": "Nom du client",
                        "type": "string"
                    },
                    "email": {
                        "description": "Email du client",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}